+++ src/AzzyBot.Bot/Commands/CoreCommands.cs
@@ -35,7 +35,7 @@ public sealed class CoreCommands
         private readonly DbActions _dbActions = dbActions;
         private readonly DiscordBotService _botService = botService;
 
-        [Command("force-channel-permissions-check"), Description("Forces a check of the permissions for the bot in the necessary channel.")]
+        [Command("force-channel-permissions-check"), Description("Forces a check of the permissions for the bot in the necessary channel."), RequirePermissions(DiscordPermissions.None, DiscordPermissions.ManageChannels)]
         public async ValueTask ForceChannelPermissionsCheckAsync(SlashCommandContext context)
         {
             ArgumentNullException.ThrowIfNull(context, nameof(context));